requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.6
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
flake8==6.1.0
flake8-django==1.4
isort